Refactors some of the walk2 code
Moves some of the walk2 data structures into walk2.h
